I CAN GRAPH STUFF (not a lot of stuff though).

+++ src/taylor/math/Constant.java
@@ -2,14 +2,14 @@ package taylor.math;
 
 public class Constant extends Expression {
     
-    public int value;
+    public double value;
     
-    public Constant(int newVal) {
+    public Constant(double newVal) {
         value = newVal;
     }
     
     @Override
-    public int getY(int x) { return value; }
+    public double getY(double x) { return value; }
     
     @Override
     public Expression derive() {
@@ -18,6 +18,6 @@ public class Constant extends Expression {
     
     @Override
     public String toString() {
-        return Integer.toString(value);
+        return Double.toString(value);
     }
 }
