+++ src/taylor/math/Expression.java
@@ -3,5 +3,5 @@ package taylor.math;
 public abstract class Expression {
     
     public abstract Expression derive();
-    public abstract int getY(int x);
+    public abstract double getY(double x);
 }
